FIFO Buffers: Refactoring and fixes
The regular FIFO Buffer is now mostly functional, except for the offset
GUI slot stuff, that is still broken.

Also note, that for example a Hopper refuses to input items, if the
wrapped FIFO buffer slot has a non-matching item in it (it doesn't try to
insert anyway). To fix that, the block most likely needs to switch to
using separate external inventories, depending on the block face.

The Pulsed FIFO Buffer is also currently still broken (it doesn't advance
on redstone pulse).