FIFO Buffers: Change wrapper to 2 slots
By using a wrapper with two slots, the second of which is always reported
to be empty, it allows things to always try to insert to the "second slot".
And since all insertions use the internal insertion position anyway, this
is a nice workaround without having to use different inventories on
different block faces to allow insertion although the wrapper slot seemed
to be full.