Pipes: New client sync packets, and TESR improvements
The amount of packets has dropped to half (more or less),
and the size of packets has also been reduced on top of that.
There is no longer a packet for every slot change to sync the item.
Instead, there are packets which tell the client side pipe TEs to move
items to/from adjacent pipes. Only new inserted items are synced to the
client once, when they first enter a pipe.

The rendering also now happens to the correct output side.
To achieve this, the items are now visually half a block behind
where they actually are.