Add a Strict Sequential Filter *WIP*
It's almost funtional, but there is still an item loss bug when the block
is broken at some specific times, like when it's outputting a filtered
item sequence.

Also, the current code (using indices and not an actual buffer inventory)
is rather messy, and will be converted to using a buffer inventory, like
the Switchable Splitter does.